pyproject.toml
[project]
name = "todoist"
version = "0.1.0"
description = "A fastapi application for managing todos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
]
